context_search_engine/requirements.txt
# Ollama Integration
ollama>=0.3.0
requests>=2.31.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
aiofiles>=23.0.0
asyncio-throttle>=1.0.0
aiohttp>=3.8.0

# Text Processing
regex>=2023.0.0
langdetect>=1.0.9

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and Monitoring
structlog>=23.0.0
prometheus-client>=0.17.0